# docs/api/web-channel-openapi.yaml
openapi: 3.1.0
info:
  title: Yodoca Web Channel API
  version: 0.4.0
  description: >
    OpenAPI specification for the experimental web channel API with first-class
    thread and project resources.
servers:
  - url: http://127.0.0.1:8080
    description: Local default
tags:
  - name: openai-compatible
  - name: ag-ui
  - name: system
  - name: threads
  - name: projects
  - name: notifications
paths:
  /v1/models:
    get:
      tags: [openai-compatible]
      summary: List available models
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OpenAI-compatible model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/chat/completions:
    post:
      tags: [openai-compatible]
      summary: Create chat completion
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionsRequest"
      responses:
        "200":
          description: Completion response (JSON or SSE stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionsResponse"
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BusyError"
        "504":
          $ref: "#/components/responses/TimeoutError"
  /v1/responses:
    post:
      tags: [openai-compatible]
      summary: Create response
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponsesRequest"
      responses:
        "200":
          description: Responses API result (JSON or SSE stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponsesResponse"
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BusyError"
        "504":
          $ref: "#/components/responses/TimeoutError"
  /agent:
    post:
      tags: [ag-ui]
      summary: Run agent (AG-UI protocol)
      description: >
        AG-UI (Agent-User Interaction) protocol endpoint. Accepts RunAgentInput
        and streams AG-UI events over Server-Sent Events. See
        https://docs.ag-ui.com/introduction for protocol details.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgUIRunRequest"
      responses:
        "200":
          description: AG-UI event stream (SSE)
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events with AG-UI event types
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BusyError"
        "504":
          $ref: "#/components/responses/TimeoutError"
  /api/health:
    get:
      tags: [system]
      summary: Health check
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/threads:
    get:
      tags: [threads]
      summary: List threads
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thread list
          content:
            application/json:
              schema:
                type: object
                properties:
                  threads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Thread"
                required: [threads]
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags: [threads]
      summary: Create thread
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadRequest"
      responses:
        "200":
          description: Created thread
          content:
            application/json:
              schema:
                type: object
                properties:
                  thread:
                    $ref: "#/components/schemas/Thread"
                required: [thread]
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/threads/{thread_id}:
    get:
      tags: [threads]
      summary: Get thread metadata and history
      security:
        - bearerAuth: []
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thread metadata and history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetailResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      tags: [threads]
      summary: Update thread metadata
      description: >
        `is_archived=true` hides the thread from default listings without
        deleting `agent_messages`. Restore with `is_archived=false`.
      security:
        - bearerAuth: []
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThreadRequest"
      responses:
        "200":
          description: Updated thread
          content:
            application/json:
              schema:
                type: object
                properties:
                  thread:
                    $ref: "#/components/schemas/Thread"
                required: [thread]
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      tags: [threads]
      summary: Archive thread
      description: >
        Soft-archives the thread. History is preserved and can be restored with
        `PATCH /api/threads/{thread_id}` and `is_archived=false`.
      security:
        - bearerAuth: []
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archive result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/projects:
    get:
      tags: [projects]
      summary: List projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                required: [projects]
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags: [projects]
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Created project
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                required: [project]
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/projects/{project_id}:
    get:
      tags: [projects]
      summary: Get project
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                required: [project]
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      tags: [projects]
      summary: Update project
      description: >
        Replaces `files` when the field is provided. If no fields are supplied,
        the project is returned unchanged and `updated_at` is not modified.
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                required: [project]
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      tags: [projects]
      summary: Delete project
      description: >
        Deletes the project and unlinks bound threads via the
        `threads.project_id` foreign key.
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/notifications:
    get:
      tags: [notifications]
      summary: Long-poll proactive notifications
      security:
        - bearerAuth: []
      parameters:
        - name: timeout
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 60
            default: 30
      responses:
        "200":
          description: Notification batch (possibly empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
  parameters:
    XThreadId:
      name: X-Thread-Id
      in: header
      required: false
      description: Optional thread identifier for routing into an existing conversation thread.
      schema:
        type: string
  responses:
    UnauthorizedError:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BusyError:
      description: Another request is currently in progress
      headers:
        Retry-After:
          description: Recommended retry delay in seconds
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TimeoutError:
      description: The request exceeded `request_timeout_seconds`
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ModelsResponse:
      type: object
      properties:
        object:
          type: string
          const: list
        data:
          type: array
          items:
            $ref: "#/components/schemas/ModelObject"
      required: [object, data]
    ModelObject:
      type: object
      properties:
        id: { type: string }
        object: { type: string, const: model }
        created: { type: integer }
        owned_by: { type: string }
      required: [id, object, created, owned_by]
    ChatCompletionsRequest:
      type: object
      properties:
        model:
          type: string
          default: yodoca
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
          default: false
        temperature:
          type: number
          minimum: 0
          maximum: 2
      required: [model, messages]
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                additionalProperties: true
      required: [role, content]
    ChatCompletionsResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, const: chat.completion }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            additionalProperties: true
        usage:
          $ref: "#/components/schemas/UsageChat"
      required: [id, object, created, model, choices, usage]
    ResponsesRequest:
      type: object
      properties:
        model:
          type: string
          default: yodoca
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                additionalProperties: true
          nullable: true
        messages:
          type: array
          nullable: true
          items:
            type: object
            additionalProperties: true
        stream:
          type: boolean
          default: false
      required: [model]
    ResponsesResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, const: response }
        status: { type: string }
        output:
          type: array
          items:
            type: object
            additionalProperties: true
        model: { type: string }
        usage:
          $ref: "#/components/schemas/UsageResponses"
      required: [id, object, status, output, model, usage]
    UsageChat:
      type: object
      properties:
        prompt_tokens: { type: integer, minimum: 0 }
        completion_tokens: { type: integer, minimum: 0 }
        total_tokens: { type: integer, minimum: 0 }
      required: [prompt_tokens, completion_tokens, total_tokens]
    UsageResponses:
      type: object
      properties:
        input_tokens: { type: integer, minimum: 0 }
        output_tokens: { type: integer, minimum: 0 }
        total_tokens: { type: integer, minimum: 0 }
      required: [input_tokens, output_tokens, total_tokens]
    HealthResponse:
      type: object
      properties:
        status: { type: string, enum: [ok] }
        uptime_seconds: { type: integer, minimum: 0 }
      required: [status, uptime_seconds]
    Thread:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string, nullable: true }
        title: { type: string, nullable: true }
        channel_id: { type: string }
        created_at: { type: integer }
        last_active_at: { type: integer }
        is_archived: { type: boolean }
      required: [id, channel_id, created_at, last_active_at, is_archived]
    ThreadDetailResponse:
      type: object
      properties:
        thread:
          $ref: "#/components/schemas/Thread"
        history:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [thread, history]
    CreateThreadRequest:
      type: object
      properties:
        id: { type: string, nullable: true }
        project_id: { type: string, nullable: true }
        title: { type: string, nullable: true }
    UpdateThreadRequest:
      type: object
      properties:
        title: { type: string, nullable: true }
        project_id: { type: string, nullable: true }
        is_archived: { type: boolean, nullable: true }
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        instructions: { type: string, nullable: true }
        agent_config:
          type: object
          additionalProperties: true
        created_at: { type: integer }
        updated_at: { type: integer }
        files:
          type: array
          items: { type: string }
      required: [id, name, agent_config, created_at, updated_at, files]
    CreateProjectRequest:
      type: object
      properties:
        name: { type: string }
        instructions: { type: string, nullable: true }
        agent_config:
          type: object
          additionalProperties: true
          nullable: true
        files:
          type: array
          items: { type: string }
      required: [name]
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string, nullable: true }
        instructions: { type: string, nullable: true }
        agent_config:
          type: object
          additionalProperties: true
          nullable: true
        files:
          type: array
          items: { type: string }
          nullable: true
    Notification:
      type: object
      properties:
        id: { type: string }
        text: { type: string }
        created_at: { type: integer }
      required: [id, text, created_at]
    NotificationsResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
      required: [notifications]
    OperationResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string, nullable: true }
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: string, nullable: true }
          required: [message, type]
      required: [error]
    AgUIRunRequest:
      type: object
      description: AG-UI RunAgentInput request body (camelCase)
      properties:
        threadId:
          type: string
          description: Conversation thread identifier (maps to thread_id)
        runId:
          type: string
          description: Unique ID for this agent run
        parentRunId:
          type: string
          nullable: true
          description: Parent run ID for branching
        messages:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              role:
                type: string
                enum: [user, assistant, system, tool, activity, reasoning, developer]
              content:
                oneOf:
                  - type: string
                  - type: array
                    items:
                      type: object
                      properties:
                        type: { type: string }
                        text: { type: string }
            additionalProperties: true
          default: []
        tools:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
              parameters: {}
            additionalProperties: true
          default: []
        context:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              value: { type: string }
            additionalProperties: true
          default: []
        state: {}
        forwardedProps: {}
      required: [threadId, runId]